refactor: 공통 버튼 컴포넌트 리팩터링 및 명칭 변경
- 용도별 구분을 위해 `AddCircleButton`을 `PlusBadgeButton`으로 변경하고 `core/ui/button` 패키지로 이동
- `PlusFAB`을 `AfternoteFloatingActionButton`으로 변경하여 `core/ui/button` 패키지로 통합 관리
- `AfternoteFloatingActionButton`의 기본 아이콘 크기를 20dp에서 24dp로 변경
- `ProfileImage`, `DiaryScreen`, `AfternoteEditor` 등 프로젝트 전반의 호출부를 새 컴포넌트로 교체 및 경로 업데이트
- 각 버튼 컴포넌트에 역할 및 제약 사항에 대한 KDoc 주석 추가
