+++ core/ui/src/main/kotlin/com/afternote/core/ui/button/AfternoteFloatingActionButton.kt
@@ -1,4 +1,4 @@
-package com.afternote.core.ui.scaffold.FAB
+package com.afternote.core.ui.button
 
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.CircleShape
@@ -15,8 +15,13 @@ import com.afternote.core.ui.R
 import com.afternote.core.ui.theme.AfternoteDesign
 import com.afternote.core.ui.theme.AfternoteTheme
 
+/**
+ * 화면 우측 하단에 떠 있는 대장 버튼(FAB).
+ *
+ * 자동으로 갖습니다. 크기나 패딩을 외부에서 받지 않습니다.
+ */
 @Composable
-fun PlusFAB(
+fun AfternoteFloatingActionButton(
     onClick: () -> Unit,
     modifier: Modifier = Modifier,
 ) {
@@ -30,15 +35,15 @@ fun PlusFAB(
         Icon(
             painter = painterResource(id = R.drawable.core_ui_circle_button_plus),
             contentDescription = stringResource(R.string.core_ui_fab_content_description_add),
-            modifier = Modifier.size(20.dp),
+            modifier = Modifier.size(24.dp),
         )
     }
 }
 
 @Preview(showBackground = true)
 @Composable
-private fun PlusFABPreview() {
+private fun AfternoteFloatingActionButtonPreview() {
     AfternoteTheme {
-        PlusFAB(onClick = {})
+        AfternoteFloatingActionButton(onClick = {})
     }
 }
